Add docker config for elasticsearch and logstash
Elasticsearch is the chosen DB to store wforce reports long-term.
Reports arrive via Logstash, using the UDP input. Logstash acts
as a NamedReportSink for wforce.
Kibana support in the Docker config is onoing due to issues with
performance on MacOS. Currently the kibana config stores the json
for a bunch of pre-configured views and dashboards for wforce data.