Parse OX Mobile App device_id strings
OX Mobile Apps provide User-Agent strings and use http.
However the User-Agent string they provide are non-standard and
so we cannot use the standard uap-cpp parser for browsers. This
also means that the OX Mobile apps need to provide a different protocol
string (mobileapi) as http(s) is already used for uap-cpp/browsers.