Enable GeoIP DBs to be reloaded
Previously the GeoIP DB would be loaded and new changes would be ignored.
This behaviour is very fast and thread safe, but meant that changes to the
GeoIP DB were ignored, which is non-optimal.
This commit adds a new command - reloadGeoIPDBs() which will reload all the
initialised DBs. Lookups are now protected by a read lock, while the reload
function uses a write lock.