New Report API functionality
This compromises a python/flask-based RESTful API module, that allows clients
to query the Elasticsearch DB for login and device information. Also allows
systems to "confirm" (or the converse) that logins were valid on behalf of
end-users.

Note this does not currently include any deployment or build scripts. In production
it would run under e.g. uwsgi and nginx or similar.