Convert IPv6 mapped v4 addresses into a plain v4 ComboAddress
When receiving v4 addresses mapped into v6 addresses, previously
these would be kept as v6 addresses. However because v4 and v6
addresses are stored n statsdbs with different prefixes, this
causes problems.
For example with a v6 prefix of /64, previously the whole IPv4
address space would be treated as a single IP address.