Fix broken regression config and related test
The regression config previously used an IP key to check changing
countries. Since a single IP can only be in a single country, this
was clearly wrong. This item is actually designed to test IP prefixes
in the regression tests, so countryCount is replaced with a
straightforward count of logins by IP. The test which looks for
countryCount is modified appropriately to look for countLogins instead.