New syncDBThread function to sync a db dump to another wforce process
This is a WIP commit. Using UDP to send the messages is not reliable because the
data can be quite large, exceeding the UDP MTU and thus causing packets to be
lost. This needs to be rewritten to support TCP.

Also the DB dump currently locks the entire for the duration of the DB. This is
not ideal, but until I can find a better way it's what we have.