Refactor wforce-webserver to use an explicit queue
Previously a fixed-size queue was used as part of the ctpl thread pool library.
However this meant that the queue length was a limiting factor for the number
of connections. Additionally because the queue was a fixed size, new connections
received once the queue size was reached would just be dropped, leaving them in a zombie state.

This commit changes connection handling to use a variable length queue, however
we do check the queue size and if it reaches a configurable limit then we stop
adding connections to the queue. Such connections will get a delayed response, but
they won't enter a zombie state. This means that the queue size is a limit on
the number of active connections (i.e. connections with data that needs to be
processed), but it's not a limit on the total number of connections (active +
inactive).

Overall this means that connection handling in overload situations will be better,
leading hopefully to better client responses and less timeouts. Certainly the
zombie state should no longer exist.