Perform sibling replication in a separate thread local to each Sibling
Previously the only mechanism to send messages was the send() method,
which sent messages synchronously, blocking the calling thread. This
meant that any delays or failures would cause the report function to
hang for a long time. This commit adds a new queueMsg() method, which
adds the replication message to a queue, which is read by a thread
initialized in the constructor, which performs the actual send.
The asynchronous queueMsg() method is now used exclusively instead of send.
The send() method is still thread-safe, in case any new code decides to
use the synchronous method.