Replication fixes (#289)
* Move Sibling class and wforce replication into separate source files

Previously Sibling and all replication code was all part of wforce.cc/hh
which was very confusing, with no clear modularization. This improves
the situation somewhat, and also allows e.g. Sibling to be used without
including all the wforce includes etc.

* Optimise Sibling queue push by calling notify outside of lock

* Remove redundant Sibling instance in syncDBThread

* Terminate queue thread in Sibling destructor

Previously the queue thread would be in a joinable state meaning
that std::terminate would be called when the thread destructor was
called. Prior to that, the thread was in a detached state, which
is also bad because the class instance would be removed without
the thread terminating.

* Don't add siblings or report sinks in the client

* Fix review nits