Make twmap expiry/expunge configurable and more frequent
Previously the expiry (which really should be called expunge/evict)
code ran at a minimum of once every 30 seconds. However that
causes problems when the stats db reaches the maximum size and the
process is under heavy load, as it has a lot of work to do, which
can take a long time, during which the DB is locked.
This commit makes two main changes:
1) Reduces the default sleep from 30 seconds to 250ms.
2) Makes the sleep configurable via a new method