Fix control socket leak when client closes connection immediately
The socket is wrapped in a Socket class, which closes the socket
when destructed. However the Socket was not declared first, which
meant that the read/write before it was declared could trigger
an exception, meaning the socket would not be closed.
Declaring the Socket class as the first line in the function
fixes this error.

(cherry picked from commit 645d0931b9793efca8a5e876f9365fd6b0fec8c7)