New functions dumpDBToNetwork(), dumpEntriesThread() and dumpEntriesToNetwork()
dumpDBToNetwork() is taken from syncDBThread() and is simply to
provide more flexibility in case we ever want
to dump the stats dbs to a socket outside of syncDBThread().

dumpEntriesThread() and dumpEntriestoNetwork() are used to send
a summary of the DB entries to a remote server. This could be
for debugging purposes, or to analyze the data stored in the
stats DB.