Use the Json version of runHook instead of the string version
In order to support Kafka REST Proxy, we use the kafka-aware version
of the runHook method. This entails using the json version of
runHook rather than the string version.

Using the json version is less efficient when there are multiple
webhooks defined for an event, because the Json must be serialised
to a string multiple times rather than once. However it provides
the required flexibility to send to kafka using the correct
json wrapper.