Fix crash bug due to race condition in Sibling class
The socket used by the Sibling class used to be initialised in the constructor
directly, by calling connectSibling().
It was then moved into a thread that itself was initialized in
the constructor. This created a possible race condition whereby the unique
pointer to the socket might not have been initialised before being referenced.
This PR fixes that by moving the connectSibling() call back into the
constructor directly.