Refactor lustate handling to be much more efficient
Previously luastates were stored in an array. Worker threads
accessed luastates by simply retrieving the "next" state in the
array, regardless of whether that state was in use or not.
This commit instead uses a pool approach, where worker threads
pop the state off the aray while the state is in use, then return
the state once it's no longer needed (automatically via destructor).
A read-only copy of the entire array is maintained for the console/
control thread which needs to be able to call all luastates.